Create InfixToPostfixUsingSinglyLinkedList.c
Recall the concepts of stacks. Implement a stack using linked list. Build a program that takes an infix statement and converts it to a postfix statement. If the statement is incorrect the program should inform the user that the input statement is incorrect. Otherwise it should return the postfix expression. After the statement has been successfully converted, the program should then evaluate this postfix expression and return the result.